requirements.txt
pandas
numpy
xlsxwriter
click
tabulate
pyyaml
# written by our group
vienna
rna_seq_tools